fix(tui): inline RevertTurnSelectorComponent into SelectorController
A standalone component file trips a tsgo inference artifact in an
unrelated test (eval-code-mode-declarations Map overload fails whenever
this file joins the program). Inlining the 40-line component into its
owning controller restores a clean workspace-wide typecheck. Also picks
up biome formatting on the touched files.
